(Temporarily?) allow XSLT <run> by itself
To accommodate TAPAS generic and TEIBP, which don't have schemas
(in the config file, <constraint>s), allow a single XSLT <run> in
the same place where an XProc <run> can appear.

This commit may be undone if it's decided that those view packages
will use XProc instead, or if some other workaround would be
better.