Create a fragment out of the TEIBP full HTML output
TAPAS-xq used to take the output of a TEIBP transformation and
make an HTML fragment out of the contents of <body>. That code is
now moved to teibp-fragment.xsl so that TAPAS-xq can handle the
view packages dynamically.

Also, xml-to-string.xsl has been updated to reflect commit 99e0289
in TAPAS-xq.