Show a profile of the boxed element clicked
When a div-like element is clicked, a section of the control
panel shows data about that element. Currently the data is in HTML,
as it is taken directly from the data attributes with no munging.
The next step is get or make this information readable.

Attributes with a prefix will now show the prefix in the name of
the corresponding HTML data attribute.