Create human-readable property list for boxed elements
The original TEI attribute names are captured in a new HTML data
attribute, and used to obtain the values of the attributes as they
appeared in the TEI document. This is slightly more complicated/
fragile, but allows the JS to display the `@ATT-NAME` version of a
TEI attribute, rather than the camel-cased name of its
corresponding HTML data attribute.