Add wrapper around counting robot <fieldset>
In Chrome, the scrollbar includes the <legend>, and consequently
poked out of the component box. There is now a wrapper div around
the entire fieldset to accommodate the scrollbar, and a @tabindex
has been added to the first <input>.

Selecting a counting robot <input> no longer scrolls one to the
first instance.