Fix element glosses; move control box in DOM
When called in eXist 2.2, the ODD interpreter glossed elements
but prefixed the glosses with the contents of any attributes on
the <gloss>. This has been fixed by obtaining the text()
descendants of the matched <gloss>.

The control box is now placed between the TEI header and the text
containers. Form fields are disabled until the Javascript enables
them. If there is no JS or CSS, the HTML output should now feel
somewhat more cohesive.