Remove element glossing by imported XSLT
When Hieractivity is used in the eXist environment, the TEI ODD
(stored as an XML document separate from the XSLT) incurs a heavy
processing cost, especially if there are multiple simultaneous
requests. To resolve this, the imported gloss-gi template has been
replaced with a local glossable-gi template, which outputs the
element name and includes data attributes to allow later glossing
by Javascript.