Create family of name-like TEI elements
Families of elements are classed using a very simple system, where
depth is only used to determine if the current element occurs
within an element of the same family, and so will need to be shown
with a variant color.

The margins of boxed elements have been reduced in 'depth-wise'
mode, and those margins stay the same in 'familial' mode. This
should hopefully reduce confusion when toggling between modes.
Boxes are now shown in familial mode, using a foggy white to
de-emphasize depth without completely taking it out.