Add specialized formatting for Hieractivity outside TAPAS reader
If a full HTML page is requested, the Hieractivity XSLT adds a
class 'standalone-html' to the <body> element. This class is used
for some minor CSS rules specific to the non-TAPAS version. More
importantly, it allows the Javascript to calculate box heights
using .outerHeight() when in the TAPAS reader, and using .height()
when in standalone mode. In other words, the boxes should look
similar in both versions.